+++ dwt/include/dwt/widgets/ScrolledContainer.h
@@ -84,10 +84,12 @@ private:
 	double getAccessibleViewSize(int type) const;
 	bool isAccessibleScrollable(int type) const;
 	void setAccessibleScrollPosition(int type, int position);
+	bool handleKeyDown(int key);
 };
 
 inline ScrolledContainer::Seed::Seed(DWORD style, DWORD exStyle) :
-BaseType::Seed(style | WS_CHILD | WS_HSCROLL | WS_VSCROLL, exStyle | WS_EX_CONTROLPARENT)
+BaseType::Seed(style | WS_CHILD | WS_TABSTOP | WS_HSCROLL | WS_VSCROLL,
+	exStyle | WS_EX_CONTROLPARENT)
 {
 }